schemeshard, drop-extsubdomain: allow nodes to reconnect during operation
Allow nodes to reconnect to the database during drop-extsubdomain operation
by fixing the moment when database path is considered deleted.

Database root path is marked as _being deleted_ at the operation start,
but it was marked actually _deleted_ also at the start of the operation.
Which is too early because path that have the  _deleted_ mark becomes
unresolvable to any external observer such as compute node that is trying
to reconnect to the database.

In the case of a dedicated database when its hive tablet has no other place to
run but on the dedicated database nodes, losing all database nodes would
prevent drop-extsubdomain operation from any chance to complete ever --
-- hive is used by operation to stop all database tablets, it has no
place to run if all database nodes had disconnected but no node is able
to reconnect again.

Databases that hosts serverless databases also suffer from this,
they are just much less likely to be removed.