tx_proxy, tests: remove excess waiting in CheckTableBecome*()
Use TTestActorRuntimeBase::DispatchTimeout manipulation to achieve
proper (small) wait times.

The reason to this is that timeout value given to CheckTableBecome*()
propagates first to Tests:TClient::WaitForTablet*(), then propagates
to NActors::TTestActorsRuntimebase::GrabEdgeEvent().

Timeout parameter to GrabEdgeEvent*() is a timeout in a simulated time.
Timeout mechanics does not work if time does not get updated during a simulation.
And entire time simulation mechanics does not work if TTestActorRuntimeBase
is used with UseRealThreads setting set to true (which is the default).

And when timeout in a simulated time does not work, the only timeout
GrabEdgeEvent*() really obeys is DispatchTimeout which defaults to 60
seconds (or 120 seconds if code runs under some sanitizer).

Tests in ydb/core/tx/tx_proxy runs with UseRealThreads set to true,
so timeouts in CheckTableBecome*() methods do nothing, and whenever
tests wait for something that should never happen (and specify small
reasonable timeout value for that) they wait for 60 (or 120) seconds instead.

This change cuts run time of the affected tests from 7 minutes to 2.