fix group usage when changing channel params via CreateTablet
The old version updated channel binding info before releasing the AU. If the storage pool was changed, this led to the old storage pool retaining outdated resource values. If the storage pool remained the same, this led to incorrect value being subtracted from storage group's resources, and potentially to underflow.