consider tablets alive only after they signal active
Local no longer counts tablets as alive until they connect to pipe and actually start operating. This way, if a tablet is stuck reading its state on restart, it will not be incorrectly reported as alive.

This is achieved by adding a new event, TEvReady, that is similar to previously used TEvRestored, but is only sent when the tablet receives TEvTabletActive from itself. This way, the only change in general tablet behavior is the addition of a new event, that may need to be handled.