Consider Sdbus-C++-LGPL-exception-1.1 as REQUIRE_DISCLOSURE_STATIC
```
As an additional permission to the GNU Lesser General Public License version
2.1, the object code form of a "work that uses the Library" may incorporate
material from a header file that is part of the Library.  You may distribute
such object code under terms of your choice, provided that:
    (i)   the header files of the Library have not been modified; and
    (ii)  the incorporated material is limited to numerical parameters, data
          structure layouts, accessors, macros, inline functions and
          templates; and
    (iii) you comply with the terms of Section 6 of the GNU Lesser General
          Public License version 2.1.
```

`contrib/libs/sdbus-cpp` is the only library with such exception, and it hardly can be used in header-only mode.