Enable license validation in JAVA_PROGRAM
EPL license family allows larger work to be distributed under arbitrary license as long as the licensed component is kept in separate file and its modifications are disclosed under the terms of EPL license.

Relevant texts are (citing [this](https://www.eclipse.org/legal/epl-v10.html) version of the license):

```
"Contribution" means:

a) in the case of the initial Contributor, the initial code and documentation distributed under this Agreement, and
b) in the case of each subsequent Contributor:
i) changes to the Program, and
ii) additions to the Program;

"Program" means the Contributions distributed in accordance with this Agreement.
...

Contributions do not include additions to the Program which: (i) are separate modules of software
distributed in conjunction with the Program under their own license agreement, and (ii) are not derivative works of the Program.
...

A Contributor may choose to distribute the Program in object code form under its own license agreement, provided that:

iv) states that source code for the Program is available from such Contributor,
    and informs licensees how to obtain it in a reasonable manner on or through a medium customarily used for software exchange.
```