CREATE / DROP VIEW at SchemeShard level
CREATE and DROP operations for the new object in SchemeShard. (ALTER is not and probably will not be implemented in the first release version of the view.) For the most part the implementation is as simple as possible and copies MkDir / RmDir, because there is no logic at SchemeShard level for non-materialised views.