Enable SELECT from views
Selecting from a view currently works for the following queries written in the view:
- SELECT 1
- SELECT * FROM SomeTable
- SELECT a, b FROM SomeTable (you can specify which columns you need)
- SELECT * FROM FirstTable JOIN SecondTable ON ...
- SELECT * FROM FirstTable UNION SELECT * FROM SecondTable
- SELECT * FROM SomeOtherView
- SELECT * FROM FirstView JOIN SecondView ON ...

The idea of the implementation is the following: whenever we encounter TExprNode corresponding to a Right! read from a view, we change this node to a TExprNode, which corresponds to the compiled query, stored in the view. This is done on RewriteIO stage of the pipeline. The biggest challenge was to meet all the expectations of the CheckTx function applied to the rewritten query.