YQL-18027: Add tests for block implementations
This patch was intended to implement the tests for block implementations
of the UDFs tested within StreamFormat.sql file. However, several
expressions used in the aforementioned file breaks vectorization
pipeline (e.g. implicit conversions), so the default input is extended
by big unsigned and small negative integer columns. Hence, to implement
the same tests for both scalar and block implementations old tests also
use the new input as a result of the patch.