Implement TGenericKernelExec
This patch implements the template of generic kernel for block
computations. It allows to implement block processing for UDFs with more
than two arguments. Process routines for such kernels recieves a tuple
with K BlockItem elements, so the particular routine implementation is
responsible for arguments unpacking.

The template kernel function implementing the batch processing makes
"dynamic" argument specialization to reduce computational pressure by
preallocaiing the tuple (i.e. vector of items) for passing the arguments
to the Process routine and hoisting the initialization of its scalar
values subset from the main processing loop.