Handle optional arguments in Arrow UDF signatures
In scope of this patch, PrepareSimpleArrowUdf helper is adjusted to
provide the support for the UDF's optional arguments. Apart from
relaxing the couple of assertions, there are two main changes made.

Unlikely to the required arguments, the type of the omitted argument has
to be passed to the specialized UDF signature builder as an original
TOptional parameter type. At the same time, all the omitted parameter
types require to be wrapped with the Block type with the Scalar shape in
the UDFKernel signature).

Finally, all the omitted arguments are substituted with the Null datum
value to be passed to the fixed-arg kernel in scope of the Run routune.