Handle optionals in SimpleSignatureType builder
Before the patch, callableBuilder instance was generated without
optional arguments even though the callable to be built via the
particular FunctionTypeInfoBuilder specified the amount of its
optionals.

Though the information about the optional parameters can be retrieved
from the given FunctionTypeInfoBuilder, it was decided to preserve
builder routine being pure functions, so its signature is extended with
the parameter to obtain the number of the optional arguments for the
callable.

This change is used in the next patch to generate Arrow UDF with the
optional parameters.