YQL-18051: Implement computation phase for BlockExists
This patch implements the helper emitting the relevant TBlockFuncNode.
There is optimization in kernel for the cases when all values in the
batch and are invalid (i.e. NULL) or valid (i.e. NOT NULL). In this case
the data buffer in the resulting array is filled by 0 or 1 respectively.