Fix offset processing in DecompressToSparseBitmap
There is a bug in the aforementioned routine, when the given offset is a
multiple of 8 and the head of the batch to be skipped is processed by
the main loop. Unfortunately, the test input uses only odd slots for
non-null data, hence the resulting output verification covered the bug.
The randomization introduced to the test in the previous commit allows
to finally reproduce and fix the misbehaviour.

Follows up #3400