Remove unnecessary abort when periodic table stats are received from "deleted" shard (#3567)
It turned out to be very possible for SchemeShard to receive TEvPeriodicTableStats from a datashard that has been removed (i.e. ShardInfos has been erased by the shardIdx). TabletIdToShardIdx entry for a shard is not erased, when the shard is being deleted. Moreover, it seems that entries in TabletIdToShardIdx are never erased, unless by .clear(), when the SchemeShard dies.