ydb(d) cli: add commands for setting interrupt-inheritance flag in acl
InterruptInheritance is a special flag that blocks permission
inheritance for the target object. There are exist a variance of
inheritance type modifiers which effect how inheritance works
for an individual access control entry, but InterruptInheritance flag
have a global effect on an object.
InterruptInheritance was part of the public api for a while now, but
there was no cli commands that could manipulate it.

Add {clear,set}-inheritance commands to `ydb` and `ydbd`.