Make CollectBlockRewrites handle Struct members
<AsStruct> arguments (i.e. the members of the resulting structure
object) are literal tuples, that don't propagate their child rewrites.
As a result, if there is a block rewrite occurred in the expression
graph below any of these lists, it takes no effect above this "list
barrier". Hence, all <AsStruct> children should be traversed separately
to either wrap the complete expressions, supported by the block engine,
with <AsScalar> callable, or apply the necessary rewrites. If none of
the conditions above are met, the block rewrite for <AsStruct> node is
aborted, since one of the children being proceseed is rewritten neither
to block, nor scalar, ergo, violating <BlockAsStruct> type requirements.