YQL-17115: Make ComputationPatternCache::Smoke deterministic
There is no way to accurately define how the entry's allocator obtained
the memory pages: using free ones from global page pool or directly
requested by <mmap>. At the same time, it is the total allocated bytes
(not just the number of the borrowed pages) that is a good estimate of
the memory consumed by the pattern cache entry for real life workload.

Hence, to avoid undesired cache flushes, the free pages of the entry's
allocator is released as a result of the patch.