auditlog: fix logged table path for SplitMergeTablePartitions
Common path reconstruction from WorkingDir+Name should not be applied
to SplitMergeTablePartitions operation as its TableName already holds
an absolute path.

SplitMergeTablePartitions usually does not get into the auditlog,
being internal operation in nature. But still it cound be issued
externally by some ydb cli command and in that case it must and
will be logged.