Reimplement IFunctionTypeInfoBuilder15 via NVI
All the interfaces provided by IFunctionTypeInfoBuilder15 has no
non-virtual wrappers in the public interface IFunctionTypeInfoBuilder,
and yield the instance of the particular version of the interface. This
might break the chain of hints for function type builder.

This patch reimplements IFunctionTypeInfoBuilder15 via NVI pattern
following the way all previous versions of this interface are written.