Throw user-friendly error if ToDate fails
There are several TzDatetime or TzTimestamp values, that cannot be
casted to TzDate. Consider the example below:
| select cast(TzDatetime("1970-01-01T23:59:59,Europe/Moscow") as TzDate)

There is no valid value, starting from the Unix epoch, to represent the
01/01/1970 midnight for Europe/Moscow timezone. Hence, this conversion
is impossible and will fail at runtime.

As a result of the patch, the exception raised while building the the
TzDate value is handled in UDF and turned into the user-friendly error
with some diagnostics.