msgbus cleanup: remove api method whoami
Msgbus wrappers are used for ydbd cli commands.
In order to remove them we want ydbd to reuse cli commands
from ydb cli. But ydb/public/lib/ydb_cli/commands library
as a whole has link dependencies (transitive from the workload
commands) that are incompatible with the dependencies of ydbd.
The minimal required portion was isolated from the library
to be shared between both the library/ydb and ydbd.

KIKIMR-21500