YT-20614: Change cross-cell copy format (in preparation for Sequoia)
The new format is better optimized for VectorizedRead, which is needed in Sequoia.

Previously, the output for the BeginCopy verb was a serialized subtree plus the list of opaque children. In a non-Sequoia world, that is useful and easy to work with. However, in Sequoia, a subtree might not be contained on the same cell, which makes it difficult to use the old data transfer format.
Now it's a vector of nodes with clear separation between them. This change makes it easier to work with VectorizedRead since the conversion between BeginCopy return and VectorizedRead(BeginCopy) return is very straight-forward. Additionally, this change helps with copying from Cypress to Sequoia, since now it's much easier to process each node individually (which is needed for Sequoia to work and not clog up one cell).

In the next PR, I'm planning on:
1. Making BeginCopy a read-request (removing snapshot locks). So the verb will only return meta-information about the requested nodes.
2. Adding a verb, executed before BeginCopy, where the locks would be taken and the general structure of the tree is returned.
017ec9971e8e0a611a7286ed748b6071cfc89048