TAsyncSemaphore's GetReadyEvent should really be non-cancelable
The `ReadyEvent_` may be shared between unrelated entities and cancellation of one entity should not trigger cancellation of another via the semaphore.

Specifically, the bug was found when debugging RemoteCopy operation.
RemoteCopy job has many tasks that copy different parts of different chunks.
These tasks are bounded via a common semaphore (a single semaphore for whole job).
These tasks are also time-bounded via `.WithTimeout`.
When semaphore is contended, tasks which are waiting on the semaphore fail to complete within timeout and get cancelled. And this cancellation is propagated to the semaphore, which cancels all tasks, which were waiting on this semaphore, even if they were copying completely unrelated chunks and even could copy chunks within their timeout.
3e49c1f30fd3e157196b8b8d76a06b6672780681