"fix" broken tests
This is not a real fix, it just masks the problem. View queries are not parsed correctly in the ParseViewQuery function. AST cannot be built in some cases. However, we didn't report AST building error as a failure previously, because we are using only the query text, not the AST. We have fixed the parsing in the other commit (see https://github.com/ydb-platform/ydb/pull/9860).