change some event scopes from `struct` to `namespace` (#10685)
Change event scopes from struct to namespace.

Only for those components, which events are used by schemeshard and in scheme operations.

When event classes are declared as inner classes of a `struct`, it is impossible to make forward declarations of those classes (without refactoring all those definitions in a very specific way). Changing event scope from `struct` to `namespace` enables easy forwarding of its event classes. And that is just a renaming and not a refactoring (no nontrivial changes).

Scopes changed:
* `TEvBlobDepot`
* `TEvBlockStore`
* `TEvColumnShard`
* `TEvConfigsDispatcher`
* `TEvConsole`
* `TEvDataShard`
* `TEvFileStore`
* `TEvHive`
* `TEvKesus`
* `TEvPersQueue`
* `TEvSchemeShard`
* `TEvSequenceShard`
* `TEvSubDomain`

This is part of "improve schemeshard operation build-time" effort (#10633).