Fix test & make use of import data instead export
Permission and user were not needed in the test so they've been removed.
The import FillItemProgress used export data instead of import data so
it's updated accordingly.

Also, in export case tables (items) are copied in temporary folder inside
database. If one tries to export table `MyTable`, it is copied to a temporary
folder created specifically for this export. In this folder `MyTable` is
named by its itemIdx.

├── MyTable
│
└── export-temporary-folder
    │
    └── 0 // MyTable copy

After that, data is exported from this copied table. This copy operation is
required to reduce the load on tables during exporting. In case of import,
there is no temporary folder. Data is imported directly in destination tables
and that's why the `DstPathId` field has been used of item instead of creating
path name, resolving it and getting its id on L49 and L53.