IGNIETFERRO-2114: correction of a debug check in the sorting predicate, a false positive of which leads to an abort.
Depending on the version of STL used, its settings, and the distribution of sortable elements, the std::sort algorithm allows the element to be compared with itself.
(For example, libcxx may do this).
In thath case, the sorting predicate is absolutely obliged to return the value `false`, and not lead to an abortion.