core/testlib: unify some client operation methods (#14141)
Make `NKikimr::Tests::TClient::{CreateUser,ModifyACL,Grant,GrantConnect}()` to behave the same as all other TClient operation methods -- -- return operation status and not assert for it to be a success. Intended use for plain TClient operation methods is to call them and check for the result status in the test body.

Add (some) `Test*()` variants that do additionally check for the success. Those are convenient in test context preparations.