IGNIETFERRO-1955:  do not rely on implementation-dependent allocator friendship behaviour
Declaring std::allocator as a friend was never guaranteed to work.
Even if it worked on some implementations, the standard did not promise that the implementation would not be altered in a way that would cause the code to fail.
And now, due to changes made to allocator_traits in C++20, it is just wrong.

For example, this code cannot be compiled with the modern version of libc++ STL library.
List of build errors: https://nda.ya.ru/t/ne-kn2lx7Cy2fS