Invoke TRope::Compact in OnVGetResult only when occupied memory excee… (#23517)
Invoke TRope::Compact in OnVGetResult only when occupied memory exceeds threshold

Changelog category

Performance improvement
Description for reviewers

This change prevents unnecessary TRope::Compact call by invoking it only when the occupied memory exceeds a threshold. Current code inside of the if statement is redundant as it is applying the same logic as TRope::Compact.

Co-authored-by: Anton Myagkov <antomyagkov@gmail.com>