Implement a test for splitting shards by the follower load (#26236)
Changelog entry

Add a new test to verify splitting shards by the follower load

Changelog category

Not for changelog (changelog entry is not required)

Description for reviewers

Added a new test TSchemeShardSplitByLoad::TableSplitsByFollowerLoad, which is supposed to verify that a shard splits, when one of the followers is overloaded. Since splitting by the follower load is not implemented yet, this test is essentially reversed - it verifies that no splitting happens in this case.

NOTE: Once this new feature is implemented, this test will be updated to verify the new behavior.