YQL-20496: Fix an option highlighting pattern
There was a bug with highlighting an identifier a in expression a == b
as an option. So for an option match we expect an exactly = token. So
all hints are enumerated and the solid regex is built.
commit_hash:8df3cd082a7eaa99fc06daa282ba6dd9f0798f6f